backend_project/requirements.txt
django==2.2.17
requests
django-cors-headers
djangorestframework-jwt
mysqlclient
numpy==1.19.3
xlrd==1.2.0
pandas==1.1.4
datetime
djangorestframework
markdown      
coreapi 
Pillow
drf-extensions
django-redis
django-filter
django-guardian
sentry-sdk
channels==2.4.0